Added a list to CustomerService to store accounts
Added a list to CustomerService to store accounts but now realize I
should probably create two separate classes. One for the account and one
for the customer. Then maybe use a map to connect the accounts with a
customer's name. This will require me to rename a lot of classes and
objects associated with those classes so I plan to create a new branch
to try this new idea without compromising the code on this branch.