Building Program Without DB
Due to not fully understanding DataBases, having not yet covered it in
class, I decided to continue the project build using arrays in the
Driver class. This is hopefully a temporary thing and I have created a
separate branch to house this new implementation. The idea is to make
sure I have a working project even if I lose points for not connecting
to a database. I'm hoping not to merge this instance of the project with
main and to instead develop a project with a database. But this is a
failsafe.